Move spin related functions to spice.spin module Move spin related tests to spice.test_spin module Add function for getting all spin table data for queried times Add test coverage for new spin table query funcition
